+++ nodalia-cards.manifest.js
@@ -1,10 +1,10 @@
 export default {
-  "pkgVersion": "1.1.0-alpha.11",
-  "contentSha256_12": "a9ec6df72459",
+  "pkgVersion": "1.1.0-alpha.13",
+  "contentSha256_12": "37ffb8af4070",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
-  "hacsFile": "nodalia-cards-1.1.0-alpha.11.js"
+  "hacsFile": "nodalia-cards-1.1.0-alpha.13.js"
 };
-export const pkgVersion = "1.1.0-alpha.11";
-export const contentSha256_12 = "a9ec6df72459";
+export const pkgVersion = "1.1.0-alpha.13";
+export const contentSha256_12 = "37ffb8af4070";
 export const file = "nodalia-cards.bundle.js";